+++ packages/apps/session-storage/shopify-app-session-storage-dynamodb/rollup.config.js
@@ -2,6 +2,6 @@ import {getConfig} from '../../../../config/rollup/rollup-utils';
 
 import * as pkg from './package.json';
 
-const config = getConfig(pkg, 'src/dynamodb.ts');
+const config = getConfig({pkg, input: 'src/dynamodb.ts'});
 
 export default config;