+++ packages/apps/session-storage/shopify-app-session-storage-test-utils/rollup.config.js
@@ -2,6 +2,6 @@ import {getConfig} from '../../../../config/rollup/rollup-utils';
 
 import * as pkg from './package.json';
 
-const config = getConfig(pkg);
+const config = getConfig({pkg});
 
 export default config;