+++ packages/apps/shopify-api/adapters/node/__tests__/node.test.ts
@@ -12,19 +12,15 @@ const nodeEnvironment: E2eTestEnvironment = {
   name: 'NodeJS',
   domain: `http://localhost:${nodeAppPort}`,
   dummyServerPort,
-  process: spawn(
-    'yarn',
-    ['node', 'adapters/__e2etests__/test_apps/test-node-app.js'],
-    {
-      env: {
-        ...process.env, // eslint-disable-line no-process-env
-        PORT: nodeAppPort,
-        HTTP_SERVER_PORT: dummyServerPort,
-      },
-      detached: true,
-      // stdio: 'inherit',
+  process: spawn('yarn', ['node', 'bundle/test-node-app.mjs'], {
+    env: {
+      ...process.env, // eslint-disable-line no-process-env
+      PORT: nodeAppPort,
+      HTTP_SERVER_PORT: dummyServerPort,
     },
-  ),
+    detached: true,
+    // stdio: 'inherit',
+  }),
   testable: true,
   ready: false,
 };