+++ packages/apps/shopify-api/adapters/web-api/__tests__/web-api.test.ts
@@ -26,7 +26,7 @@ const webApiEnvironment: E2eTestEnvironment = {
       `${webApiAppPort}`,
       '--inspector-port',
       '9259',
-      'bundle/test-web-api-app.js',
+      'bundle/test-web-api-app.mjs',
     ],
     {
       detached: true,