+++ packages/apps/shopify-api/rest/admin/2022-10/application_credit.ts
@@ -48,6 +48,7 @@ export class ApplicationCredit extends Base {
   ): Promise<ApplicationCredit | null> {
     const result = await this.baseFind<ApplicationCredit>({
       session: session,
+      requireIds: true,
       urlIds: {"id": id},
       params: {"fields": fields},
     });