+++ packages/apps/shopify-api/rest/admin/2023-04/transaction.ts
@@ -57,6 +57,7 @@ export class Transaction extends Base {
   ): Promise<Transaction | null> {
     const result = await this.baseFind<Transaction>({
       session: session,
+      requireIds: true,
       urlIds: {"id": id, "order_id": order_id},
       params: {"fields": fields, "in_shop_currency": in_shop_currency},
     });